+++ test/test_06_moap.py
@@ -1,6 +1,8 @@
 import os
 import unittest
 
+import numpy as np
+
 from gimmemotifs.maelstrom.moap import moap
 
 
@@ -8,10 +10,11 @@ class TestMoap(unittest.TestCase):
     """A test class to test Moap functionality"""
 
     def setUp(self):
+        self.random_state = np.random.RandomState(1)
         self.data_dir = "test/data/moap"
-        self.clusters = os.path.join(self.data_dir, "clusters.txt")
-        self.motifs_count = os.path.join(self.data_dir, "motifs.count.txt")
-        self.motifs_score = os.path.join(self.data_dir, "motifs.score.txt")
+        self.clusters4 = os.path.join(self.data_dir, "4clusters.txt")
+        self.motifs_count4 = os.path.join(self.data_dir, "4clusters.motifs.count.txt")
+        self.motifs_score4 = os.path.join(self.data_dir, "4clusters.motifs.score.txt")
 
         self.clusters2 = os.path.join(self.data_dir, "2clusters.txt")
         self.motifs_count2 = os.path.join(self.data_dir, "2clusters.motifs.count.txt")
@@ -20,27 +23,30 @@ class TestMoap(unittest.TestCase):
     def test1_moap(self):
         """Test motif activity prediction"""
 
+        # for method in ["bayesianridge", "mwu", "rf", "multitasklasso", "svr"]:
         for method in ["mwu", "rf"]:
             df = moap(
-                self.clusters,
+                self.clusters4,
                 method=method,
                 scoring="score",
-                motiffile=self.motifs_score,
+                motiffile=self.motifs_score4,
             )
             self.assertEquals((623, 4), df.shape)
 
+        # for method in ["hypergeom", "bayesianridge", "rf", "multitasklasso", "svr"]:
         for method in ["hypergeom"]:
             df = moap(
-                self.clusters,
+                self.clusters4,
                 method=method,
                 scoring="count",
-                motiffile=self.motifs_count,
+                motiffile=self.motifs_count4,
             )
             self.assertEquals((623, 4), df.shape)
 
     def test2_moap(self):
         """Test motif activity prediction for two clusters"""
 
+        # for method in ["bayesianridge", "mwu", "rf", "multitasklasso", "svr"]:
         for method in ["mwu", "rf"]:
             df = moap(
                 self.clusters2,
@@ -50,6 +56,7 @@ class TestMoap(unittest.TestCase):
             )
             self.assertEquals((623, 2), df.shape)
 
+        # for method in ["hypergeom", "bayesianridge", "rf", "multitasklasso", "svr"]:
         for method in ["hypergeom"]:
             df = moap(
                 self.clusters2,